Move ResultStorage to separate file
